fix: types and lang switch
