+++ packages/umi-ui/client/src/global.d.ts
@@ -1,3 +1,4 @@
+import { Context } from 'react';
 import { IContext, ILocale, IService } from './typings';
 
 declare module '@tmp/history';
@@ -8,7 +9,7 @@ declare global {
   interface Window {
     xterm: any;
     g_uiPlugins?: any[];
-    g_uiContext?: IContext;
+    g_uiContext?: Context<IContext>;
     g_uiLocales: ILocale[];
     g_service: IService;
   }
